pytorch_triage_env/server/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
openenv-core>=0.2.0
websockets>=12.0
httpx>=0.25.0
python-multipart>=0.0.6
openai>=1.0.0
pyyaml>=6.0
